Setup dev branch
Folder structure for GCP Cloud Functions
