+++ cloud/terraform/outputs.tf
@@ -1,0 +1,3 @@
+output "bucket_url" {
+  value = "gs://${google_storage_bucket.airbnb_data_bucket.name}"
+}
