+++ cloud/terraform/terraform.tfvars
@@ -1,0 +1,2 @@
+project_id = "utilitarian-web-457114-m1"
+bucket_name = "airbnb-data-raw-europe"
