+++ cloud/terraform/variables.tf
@@ -1,0 +1,9 @@
+variable "project_id" {
+        type = string
+        description = "The GCP project ID"
+}
+
+variable "bucket_name" {
+        type = string
+        description = "The name of the GCS bucket to create"
+}
