+++ on_prem/run_pipeline.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+# Activate your virtual environment (adjust path if needed)
+source /Users/jsthiara/Desktop/NYC_Airbnb_Pipeline/venv/bin/activate
+
+# Run your pipeline script
+python /Users/jsthiara/Desktop/NYC_Airbnb_Pipeline/src/main.py
