--- terraform/terraform.tfvars
@@ -1,2 +1,0 @@
-project_id = "utilitarian-web-457114-m1"
-bucket_name = "airbnb-data-raw-europe"
